FIT-14: Add unit and integration test

+++ src/main/java/com/fittracker/fittracker/repository/WeightRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 import java.time.LocalDate;
 import java.util.Optional;
 
-public interface WeightRepository extends CrudRepository<Weight,Integer> {
+public interface WeightRepository extends CrudRepository<Weight,Long> {
     Boolean existsByDate(LocalDate date);
 
     Optional<Weight> findByDate(LocalDate date);
